Move HomeInjector to navigation package
